activating/deactivating AP takes quite some time, UI is stuck during operation, watch dog may be angry. So better do blocking wifi configuration in background

+++ playground/src/device-settings/WifiSetting.cpp
@@ -4,29 +4,52 @@
 
 WifiSetting::WifiSetting(Settings& settings)
     : BooleanSetting(settings, WifiSetting::pollAccessPointRunning())
+    , m_bgThread([&]() { this->doBackgroundWork(); })
 {
 }
 
-
-auto printRet = [](SpawnCommandLine & cmd)
-{
+auto printRet = [](SpawnCommandLine& cmd) {
   DebugLevel::warning("COut:", cmd.getStdOutput(), "CError:", cmd.getStdError(), "Exit Status:", cmd.getExitStatus());
 };
 
-WifiSetting::~WifiSetting() = default;
+WifiSetting::~WifiSetting()
+{
+  m_stopThread = true;
+  m_condition.notify_all();
+
+  if(m_bgThread.joinable())
+    m_bgThread.join();
+}
 
 bool WifiSetting::set(tEnum m)
 {
   auto ret = super::set(m);
-
-  if(get()) {
-    SpawnCommandLine cmd("su - root -c \"systemctl enable accesspoint && systemctl start accesspoint\"");
-    printRet(cmd);
-  } else {
-    SpawnCommandLine cmd("su - root -c \"systemctl disable accesspoint && systemctl stop accesspoint\"");
-    printRet(cmd);
-  }
+  m_condition.notify_all();
   return ret;
+}
+
+void WifiSetting::doBackgroundWork()
+{
+  std::unique_lock<std::mutex> lock(m_mutex);
+
+  while(!m_stopThread)
+  {
+    m_condition.wait(lock);
+
+    if(m_stopThread)
+      break;
+
+    if(get())
+    {
+      SpawnCommandLine cmd("su - root -c \"systemctl enable accesspoint && systemctl start accesspoint\"");
+      printRet(cmd);
+    }
+    else
+    {
+      SpawnCommandLine cmd("su - root -c \"systemctl disable accesspoint && systemctl stop accesspoint\"");
+      printRet(cmd);
+    }
+  }
 }
 
 bool WifiSetting::pollAccessPointRunning()
